복수형으로 바꾸고 파라미터를 알기 쉽도록 변경
이전에 onModuleInit을 통해서 순환 참조 문제를 회피하려고 했더니 user에 articleId를 넘기는 과정에서 id가 제대로 넘어가지 않는 에러를 발견함. 고칠 필요가 있음

+++ api/src/articles/service/articles.service.ts
@@ -1,14 +1,14 @@
 import { Injectable } from '@nestjs/common';
 import { ModuleRef } from '@nestjs/core';
-import { UserService } from 'src/users/service/users.service';
+import { UsersService } from 'src/users/service/users.service';
 import { ArticleRepository } from '../articles.repository';
-import { Article } from '../articles.schema';
+import { Articles } from '../articles.schema';
 import { CreateArticleDto } from '../dtos/createArticle.dto';
 import { UpdateArticleDto } from '../dtos/updateArticle.dto';
 
 @Injectable()
 export class ArticlesService {
-  private userService: UserService;
+  private userService: UsersService;
 
   constructor(
     private moduleRef: ModuleRef,
@@ -17,30 +17,35 @@ export class ArticlesService {
 
   onModuleInit() {
     // onModuleInit 생명주기 훅을 사용하여 userService 인스턴스를 늦게 가져옴.
-    this.userService = this.moduleRef.get(UserService, { strict: false });
+    this.userService = this.moduleRef.get(UsersService, { strict: false });
   }
 
   // create new article
   async create(createArticleDto: CreateArticleDto, userId: string) {
     const newArticle = await this.articleRepository.create(createArticleDto);
-    await this.userService.addArticleToUser(userId, newArticle._id); // User에 Article ID 추가
-    await this.userService.plusPrivateArticle(userId);
-    await this.userService.updateAllArticles(userId);
+    if (newArticle) {
+      await this.userService.addArticleToUser(userId, newArticle._id); // User에 Article ID 추가
+      await this.userService.plusPrivateArticle(userId);
+      await this.userService.updateAllArticles(userId);
+    }
 
-    return newArticle.readOnlyData;
+    return newArticle;
   }
 
   // update (article)id's article
-  async update(id: string, updateArticleDto: UpdateArticleDto) {
-    return this.articleRepository.update(id, updateArticleDto);
+  async update(articleId: string, updateArticleDto: UpdateArticleDto) {
+    return this.articleRepository.update(articleId, updateArticleDto);
   }
 
   // delete (article)id's article
   async delete(articleId: string, userId: string) {
-    await this.userService.removeArticleFromUser(userId, articleId); // User에 Article ID 삭제
-    await this.userService.minusPrivateArticle(userId);
-    await this.userService.updateAllArticles(userId);
     const deleteArticle = await this.articleRepository.delete(articleId);
+
+    if (deleteArticle) {
+      await this.userService.removeArticleFromUser(userId, articleId); // User에 Article ID 삭제
+      await this.userService.minusPrivateArticle(userId);
+      await this.userService.updateAllArticles(userId);
+    }
 
     return deleteArticle;
   }
@@ -50,13 +55,13 @@ export class ArticlesService {
   }
 
   // show all articles
-  async findAll(): Promise<Article[]> {
+  async findAll(): Promise<Articles[]> {
     return this.articleRepository.findAll();
   }
 
   // show (article)id's article
-  async findOne(id: string) {
-    return this.articleRepository.findOne(id);
+  async findOne(articleId: string) {
+    return this.articleRepository.findOne(articleId);
   }
 
   async findArticlesIdByAuthorId(authorId: string) {
